Restructure printing code
