+++ symengine/basic.cpp
@@ -1,4 +1,4 @@
-#include <symengine/printer.h>
+#include <symengine/printers.h>
 #include <symengine/subs.h>
 
 namespace SymEngine
@@ -21,8 +21,7 @@ int Basic::__cmp__(const Basic &o) const
 
 std::string Basic::__str__() const
 {
-    StrPrinter strPrinter;
-    return strPrinter.apply(*this);
+    return str(*this);
 }
 
 RCP<const Basic> Basic::subs(const map_basic_basic &subs_dict) const
