+++ symengine/printers/strprinter.h
@@ -1,12 +1,12 @@
-#ifndef SYMENGINE_PRINTER_H
-#define SYMENGINE_PRINTER_H
+#ifndef SYMENGINE_STR_PRINTER_H
+#define SYMENGINE_STR_PRINTER_H
 
 #include <symengine/visitor.h>
+#include <symengine/printers.h>
 
 namespace SymEngine
 {
 
-std::string ascii_art();
 std::string print_double(double d);
 std::vector<std::string> init_str_printer_names();
 
@@ -18,26 +18,10 @@ private:
     PrecedenceEnum precedence;
 
 public:
-    void bvisit(const Add &x)
-    {
-        precedence = PrecedenceEnum::Add;
-    }
-
-    void bvisit(const Mul &x)
-    {
-        precedence = PrecedenceEnum::Mul;
-    }
-
-    void bvisit(const Relational &x)
-    {
-        precedence = PrecedenceEnum::Relational;
-    }
-
-    void bvisit(const Pow &x)
-    {
-        precedence = PrecedenceEnum::Pow;
-    }
-
+    void bvisit(const Add &x);
+    void bvisit(const Mul &x);
+    void bvisit(const Relational &x);
+    void bvisit(const Pow &x);
     template <typename Poly>
     void bvisit_upoly(const Poly &x)
     {
@@ -70,11 +54,7 @@ public:
         bvisit_upoly(down_cast<const Poly &>(x));
     }
 
-    void bvisit(const GaloisField &x)
-    {
-        // iterators need to be implemented
-        // bvisit_upoly(x);
-    }
+    void bvisit(const GaloisField &x);
 
     template <typename Container, typename Poly>
     void bvisit(const MSymEnginePoly<Container, Poly> &x)
@@ -104,84 +84,23 @@ public:
         }
     }
 
-    void bvisit(const Rational &x)
-    {
-        precedence = PrecedenceEnum::Add;
-    }
-
-    void bvisit(const Complex &x)
-    {
-        if (x.is_re_zero()) {
-            if (x.imaginary_ == 1) {
-                precedence = PrecedenceEnum::Atom;
-            } else {
-                precedence = PrecedenceEnum::Mul;
-            }
-        } else {
-            precedence = PrecedenceEnum::Add;
-        }
-    }
-
-    void bvisit(const Integer &x)
-    {
-        if (x.is_negative()) {
-            precedence = PrecedenceEnum::Mul;
-        } else {
-            precedence = PrecedenceEnum::Atom;
-        }
-    }
-
-    void bvisit(const RealDouble &x)
-    {
-        if (x.is_negative()) {
-            precedence = PrecedenceEnum::Mul;
-        } else {
-            precedence = PrecedenceEnum::Atom;
-        }
-    }
-
+    void bvisit(const Rational &x);
+    void bvisit(const Complex &x);
+    void bvisit(const Integer &x);
+    void bvisit(const RealDouble &x);
 #ifdef HAVE_SYMENGINE_PIRANHA
-    void bvisit(const URatPSeriesPiranha &x)
-    {
-        precedence = PrecedenceEnum::Add;
-    }
-
-    void bvisit(const UPSeriesPiranha &x)
-    {
-        precedence = PrecedenceEnum::Add;
-    }
+    void bvisit(const URatPSeriesPiranha &x);
+    void bvisit(const UPSeriesPiranha &x);
 #endif
-    void bvisit(const ComplexDouble &x)
-    {
-        precedence = PrecedenceEnum::Add;
-    }
+    void bvisit(const ComplexDouble &x);
 #ifdef HAVE_SYMENGINE_MPFR
-    void bvisit(const RealMPFR &x)
-    {
-        if (x.is_negative()) {
-            precedence = PrecedenceEnum::Mul;
-        } else {
-            precedence = PrecedenceEnum::Atom;
-        }
-    }
+    void bvisit(const RealMPFR &x);
 #endif
 #ifdef HAVE_SYMENGINE_MPC
-    void bvisit(const ComplexMPC &x)
-    {
-        precedence = PrecedenceEnum::Add;
-    }
+    void bvisit(const ComplexMPC &x);
 #endif
-
-    void bvisit(const Basic &x)
-    {
-        precedence = PrecedenceEnum::Atom;
-    }
-
-    PrecedenceEnum getPrecedence(const RCP<const Basic> &x)
-    {
-        (*x).accept(*this);
-        return precedence;
-    }
+    void bvisit(const Basic &x);
+    PrecedenceEnum getPrecedence(const RCP<const Basic> &x);
 };
 
 std::vector<std::string> init_str_printer_names();
@@ -284,4 +203,4 @@ public:
 };
 }
 
-#endif // SYMENGINE_PRINTER_H
+#endif // SYMENGINE_STR_PRINTER_H
