+++ symengine/tests/printing/test_ccode.cpp
@@ -2,7 +2,7 @@
 
 #include <symengine/logic.h>
 #include <symengine/add.h>
-#include <symengine/codegen.h>
+#include <symengine/printers/codegen.h>
 #include <symengine/sets.h>
 
 using SymEngine::Basic;
